+++ packages/vike-vue/src/components/Config/Config-client.ts
@@ -3,8 +3,8 @@ export { Config }
 // Same as ./Config-server.ts but importing useConfig-client.js
 
 import { defineComponent } from 'vue'
-import { useConfig, configsClientSide } from '../../hooks/useConfig/useConfig-client.js'
-import type { ConfigFromHook } from '../../+config.js'
+import { useConfig } from '../../hooks/useConfig/useConfig-client.js'
+import { ConfigFromHook } from '../../types/Config.js'
 
 const Config = defineComponent<ConfigFromHook>(
   (props) => {
@@ -15,6 +15,6 @@ const Config = defineComponent<ConfigFromHook>(
   // manual runtime props declaration is currently still needed.
   // see https://vuejs.org/api/general.html#function-signature
   {
-    props: [...configsClientSide],
+    props: ['title'],
   },
 )
