+++ packages/vike-vue/src/components/Config/Config-server.ts
@@ -3,8 +3,8 @@ export { Config }
 // Same as ./Config-client.ts but importing useConfig-server.js
 
 import { defineComponent } from 'vue'
-import { useConfig, configsCumulative, configsOverridable } from '../../hooks/useConfig/useConfig-server.js'
-import type { ConfigFromHook } from '../../+config.js'
+import { useConfig } from '../../hooks/useConfig/useConfig-server.js'
+import { ConfigFromHook } from '../../types/Config.js'
 
 /**
  * Set configurations inside Vue components.
@@ -24,6 +24,6 @@ const Config = defineComponent<ConfigFromHook>(
   // manual runtime props declaration is currently still needed.
   // see https://vuejs.org/api/general.html#function-signature
   {
-    props: [...configsCumulative, ...configsOverridable],
+    props: ['Head', 'description', 'favicon', 'image', 'title', 'viewport'],
   },
 )
